Resolve conflicts

+++ main/helper.py
@@ -25,7 +25,7 @@ def country_year_list(df):
     
     return years, country
 
-def fetch_medal_tally(df, year, country):
+def fetch_medal_tally(df,year, country):
     medal_df = df.drop_duplicates(subset=['Team', 'NOC', 'Games', 'Year', 'City', 'Sport', 'Event', 'Medal'])
     flag = 0
     if year == "Overall" and country == "Overall":
@@ -43,5 +43,8 @@ def fetch_medal_tally(df, year, country):
     else:
         x = temp_df.groupby('region').sum()[['Gold', 'Silver', 'Bronze']].sort_values('Gold', ascending = False).reset_index()
     x['total'] = x['Gold'] + x['Silver'] + x['Bronze']
-    
+    x['Gold'] = x['Gold'].astype('int')
+    x['Silver'] = x['Silver'].astype('int')
+    x['Bronze'] = x['Bronze'].astype('int')
+    x['total'] = x['total'].astype('int')
     return x
